ci: update metrics workflow configuration
This commit modifies the metrics.yml workflow configuration file. The changes include updating the `base_indepth` field to true and changing the value of `repositories` from 300 to 500 in the `with` block. These changes will ensure that the base repository and its forks are included in the metrics generation process. Additionally, the `config_timezone` field is set to "Asia/Tehran" to specify the timezone for the generated metrics.