ci: change the default batch size for repository fetching
Increase the batch size from 300 to 500 in order to improve performance and efficiency. This change affects multiple jobs in the metrics.yml file. The new batch size will allow fetching repositories in batches of 100, reducing the number of requests made during the process.

Note: This commit message assumes that the previous default batch size was 300.