ci: update metrics workflow to include more detailed repository data
This commit modifies the metrics workflow in order to include more detailed repository data. The change adds a new parameter `base_indepth` with the value `yes`, which enables the collection of additional repository information. The previous version of the workflow only collected data related to the activity and community of the repositories. With this update, the workflow will now collect more comprehensive data for up to 500 repositories, including information about forks.